requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
APScheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
eventlet==0.35.2
SQLAlchemy==2.0.23
alembic==1.13.0
